Moved subcomponents into sections/
