+++ src/components/page/sections/ExperienceSection.js
@@ -1,8 +1,8 @@
 import * as React from 'react'
 
-import { Banner } from '../Banner'
-import Carousel from '../Carousel'
-import ExperienceTemplate from '../ExperienceTemplate'
+import { Banner } from './Banner'
+import Carousel from './Carousel'
+import ExperienceTemplate from './ExperienceTemplate'
 import experiences from 'content/experiences'
 
 export function ExperienceSection() {
