+++ src/components/page/sections/ProjectsSections.js
@@ -1,8 +1,8 @@
 import * as React from 'react'
 
-import { Banner } from '../Banner'
-import Carousel from '../Carousel'
-import ProjectTemplate from '../ProjectTemplate'
+import { Banner } from './Banner'
+import Carousel from './Carousel'
+import ProjectTemplate from './ProjectTemplate'
 import projects from 'content/projects'
 
 export function ProjectsSection() {
